backend/requirements.txt
django==4.0.4
djangorestframework==3.13.1
django-filter==22.1
pillow==9.1.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0
sqlparse==0.3.1